backend/requirements.txt
openai
googlemaps
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain_community
mysql-connector-python
numpy
geopy
faiss-cpu
Jinja2
requests
typing
dataclasses
psycopg2-binary 
uuid